move common function between flask and starlette to more central place
